backend/requirements.txt

fastapi
uvicorn
python-multipart
pydantic
langchain
langchain-community
fastembed
faiss-cpu
google-generativeai
pdfminer.six
python-dotenv
deep-translator
python-dotenv